wapi unit test fix

+++ wallet/unittests/wallet_api_test.cpp
@@ -191,7 +191,7 @@ namespace
 
             for (int i = 0; i < Count; i++)
             {                
-                WALLET_CHECK(Coin::FromString(result[i]["id"])->m_Idx == 132 + i);
+                WALLET_CHECK(Coin::FromString(result[i]["id"])->m_Idx == uint64_t(132 + i));
                 WALLET_CHECK(result[i]["amount"] == 1234 + i);
                 WALLET_CHECK(result[i]["type"] == "norm");
                 WALLET_CHECK(result[i]["maturity"] == 60);
